added device and group command to cli
